requirements.txt
git+https://github.com/antoniojkim/Multiprocessor.git
matplotlib
numpy
opencv
pillow
